combine TamTruService and TamTruDisplayService, combine TamVangService and TamVangDisplayService, change comment
