+++ src/main/java/com/se07/controller/services/TamTruService.java
@@ -10,7 +10,15 @@ import java.sql.*;
 import java.util.Date;
 import java.util.Optional;
 
+/**
+ * class thực hiện các phương thức CRUD với tam_tru
+ */
 public class TamTruService {
+
+    /**
+     *
+     * @return tất cả bản ghi tạm trú dưới dạng TamTruModel
+     */
     public ObservableList<TamTruModel> getAllTamTru() {
         ObservableList<TamTruModel> listTamTru = FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
@@ -26,7 +34,7 @@ public class TamTruService {
                         rs.getNString("hoTen"),
                         rs.getDate("tuNgay"),
                         rs.getDate("denNgay"),
-                        rs.getString("lydo"),
+                        rs.getNString("lydo"),
                         rs.getNString("tinhTrang"),
                         rs.getInt("idNguoiThucHien"));
                 listTamTru.add(temp);
@@ -39,94 +47,157 @@ public class TamTruService {
         return listTamTru;
     }
 
-    public ObservableList<TamTruModel> getTamTruByCCCD(String CCCD) {
-        ObservableList<TamTruModel> listTamTru = FXCollections.observableArrayList();
+    /**
+     *
+     * @return tất cả bản ghi tạm trú dưới dạng  TamTruDisplayModel
+     */
+    public ObservableList<TamTruDisplayModel> getDisplayTamTru(){
+        ObservableList<TamTruDisplayModel> list= FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
-        String query = "select * from tam_tru where CCCD = '" + CCCD + "'";
+        String query= "select CCCD, tam_tru.hoTen, ho_khau.diaChi as noiTamTru, tuNgay, denNgay, lydo, tam_tru.tinhTrang, tam_tru.idNguoiThucHien" +
+                "from tam_tru, ho_khau where tam_tru.maHoKhau= ho_khau.maHoKhau";
         try {
-            Statement statement = connection.createStatement();
+            Statement statement =connection.createStatement();
             ResultSet rs = statement.executeQuery(query);
             while (rs.next()) {
-                TamTruModel temp = new TamTruModel(
-                        rs.getInt("maTamTru"),
-                        rs.getString("maHoKhau"),
+                TamTruDisplayModel temp = new TamTruDisplayModel(
                         rs.getString("CCCD"),
                         rs.getNString("hoTen"),
+                        rs.getNString("noiTamTru"),
                         rs.getDate("tuNgay"),
                         rs.getDate("denNgay"),
-                        rs.getString("lydo"),
+                        rs.getNString("lydo"),
                         rs.getNString("tinhTrang"),
                         rs.getInt("idNguoiThucHien"));
-                listTamTru.add(temp);
+
+                list.add(temp);
             }
             statement.close();
             connection.close();
-        } catch (Exception e) {
+
+        }catch (Exception e){
             e.printStackTrace();
         }
-        return listTamTru;
+        return list;
     }
 
-    public ObservableList<TamTruModel> getTamTruByHoTen(String hoTen) {
-        ObservableList<TamTruModel> listTamTru = FXCollections.observableArrayList();
+    /**
+     *
+     * @param CCCD
+     * @return thông tin tạm trú theo căn cước công dân
+     */
+    public ObservableList<TamTruDisplayModel> getDisplayTamTruByCCCD(String CCCD){
+        ObservableList<TamTruDisplayModel> list= FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
-        String query = "select * from tam_tru where hoTen like N'%" + hoTen + "%'";
+        String query= "select CCCD, tam_tru.hoTen, ho_khau.diaChi as noiTamTru, tuNgay, denNgay, lydo, tam_tru.tinhTrang, tam_tru.idNguoiThucHien" +
+                "from tam_tru, ho_khau where tam_tru.maHoKhau= ho_khau.maHoKhau and CCCD LIKE '%"+ CCCD + "%'";
         try {
-            Statement statement = connection.createStatement();
+            Statement statement =connection.createStatement();
             ResultSet rs = statement.executeQuery(query);
             while (rs.next()) {
-                TamTruModel temp = new TamTruModel(
-                        rs.getInt("maTamTru"),
-                        rs.getString("maHoKhau"),
+                TamTruDisplayModel temp = new TamTruDisplayModel(
                         rs.getString("CCCD"),
                         rs.getNString("hoTen"),
+                        rs.getNString("noiTamTru"),
                         rs.getDate("tuNgay"),
                         rs.getDate("denNgay"),
-                        rs.getString("lydo"),
+                        rs.getNString("lydo"),
                         rs.getNString("tinhTrang"),
                         rs.getInt("idNguoiThucHien"));
-                listTamTru.add(temp);
+
+                list.add(temp);
             }
             statement.close();
             connection.close();
-        } catch (Exception e) {
+
+        }catch (Exception e){
             e.printStackTrace();
         }
-        return listTamTru;
+        return list;
     }
 
-    public ObservableList<TamTruModel> getTamTruByTen(String ten) {
-        ObservableList<TamTruModel> listTamTru = FXCollections.observableArrayList();
+    /**
+     *
+     * @param hoTen
+     * @return thông tin tạm trú theo họ tên
+     */
+    public ObservableList<TamTruDisplayModel> getDisplayTamTruByHoTen(String hoTen){
+        ObservableList<TamTruDisplayModel> list= FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
-        String query = "select * from tam_tru where hoTen LIKE N'%" + ten + "%'";
+        String query= "select CCCD, tam_tru.hoTen, ho_khau.diaChi as noiTamTru, tuNgay, denNgay, lydo, tam_tru.tinhTrang, tam_tru.idNguoiThucHien" +
+                "from tam_tru, ho_khau where tam_tru.maHoKhau= ho_khau.maHoKhau and hoTen LIKE N'%"+ hoTen + "%'";
         try {
-            Statement statement = connection.createStatement();
+            Statement statement =connection.createStatement();
             ResultSet rs = statement.executeQuery(query);
             while (rs.next()) {
-                TamTruModel temp = new TamTruModel(
-                        rs.getInt("maTamTru"),
-                        rs.getString("maHoKhau"),
+                TamTruDisplayModel temp = new TamTruDisplayModel(
                         rs.getString("CCCD"),
                         rs.getNString("hoTen"),
+                        rs.getNString("noiTamTru"),
                         rs.getDate("tuNgay"),
                         rs.getDate("denNgay"),
-                        rs.getString("lydo"),
+                        rs.getNString("lydo"),
                         rs.getNString("tinhTrang"),
                         rs.getInt("idNguoiThucHien"));
-                listTamTru.add(temp);
+
+                list.add(temp);
             }
             statement.close();
             connection.close();
-        } catch (Exception e) {
+
+        }catch (Exception e){
             e.printStackTrace();
         }
-        return listTamTru;
+        return list;
     }
 
-    public ObservableList<TamTruModel> getTamTruWhereTuNgayBetween(java.util.Date low, Date high) {
-        ObservableList<TamTruModel> listTamTru = FXCollections.observableArrayList();
+    /**
+     *
+     * @param noiTamTru
+     * @return thông tin tạm trú theo nơi tạm trú
+     */
+
+    public ObservableList<TamTruDisplayModel> getDisplayTamTruByNoiTamTru(String noiTamTru){
+        ObservableList<TamTruDisplayModel> list= FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
-        String query = "select * from tam_tru where tuNgay between ? and ?";
+        String query= "select CCCD, tam_tru.hoTen, ho_khau.diaChi as noiTamTru, tuNgay, denNgay, lydo, tam_tru.tinhTrang, tam_tru.idNguoiThucHien" +
+                "from tam_tru, ho_khau where tam_tru.maHoKhau= ho_khau.maHoKhau and ho_khau.diaChi LIKE N'%"+ noiTamTru + "%'";
+        try {
+            Statement statement =connection.createStatement();
+            ResultSet rs = statement.executeQuery(query);
+            while (rs.next()) {
+                TamTruDisplayModel temp = new TamTruDisplayModel(
+                        rs.getString("CCCD"),
+                        rs.getNString("hoTen"),
+                        rs.getNString("noiTamTru"),
+                        rs.getDate("tuNgay"),
+                        rs.getDate("denNgay"),
+                        rs.getNString("lydo"),
+                        rs.getNString("tinhTrang"),
+                        rs.getInt("idNguoiThucHien"));
+
+                list.add(temp);
+            }
+            statement.close();
+            connection.close();
+
+        }catch (Exception e){
+            e.printStackTrace();
+        }
+        return list;
+    }
+
+    /**
+     *
+     * @param low
+     * @param high
+     * @return thông tin tạm trú với ngày bắt đầu tạm trú trong khoảng từ low dến high
+     */
+    public ObservableList<TamTruDisplayModel> getDisplayTamTruWhereTuNgayBetween(java.util.Date low, Date high){
+        ObservableList<TamTruDisplayModel> list= FXCollections.observableArrayList();
+        Connection connection = ConnectionDatabase.getConnection();
+        String query= "select CCCD, tam_tru.hoTen, ho_khau.diaChi as noiTamTru, tuNgay, denNgay, lydo, tam_tru.tinhTrang, tam_tru.idNguoiThucHien" +
+                "from tam_tru, ho_khau where tam_tru.maHoKhau= ho_khau.maHoKhau and tuNgay between ? and ? ";
         try {
             PreparedStatement statement = connection.prepareStatement(query);
             statement.setDate(1, new java.sql.Date(low.getTime()));
@@ -138,34 +209,43 @@ public class TamTruService {
             e.printStackTrace();
         }
         try {
-            Statement statement = connection.createStatement();
+            Statement statement =connection.createStatement();
             ResultSet rs = statement.executeQuery(query);
             while (rs.next()) {
-                TamTruModel temp = new TamTruModel(
-                        rs.getInt("maTamTru"),
-                        rs.getString("maHoKhau"),
+                TamTruDisplayModel temp = new TamTruDisplayModel(
                         rs.getString("CCCD"),
                         rs.getNString("hoTen"),
+                        rs.getNString("noiTamTru"),
                         rs.getDate("tuNgay"),
                         rs.getDate("denNgay"),
-                        rs.getString("lydo"),
+                        rs.getNString("lydo"),
                         rs.getNString("tinhTrang"),
                         rs.getInt("idNguoiThucHien"));
-                listTamTru.add(temp);
+
+                list.add(temp);
             }
             statement.close();
             connection.close();
-        } catch (Exception e) {
+
+        }catch (Exception e){
             e.printStackTrace();
         }
-
-        return listTamTru;
+        return list;
     }
 
-    public ObservableList<TamTruModel> getTamTruWhereDenNgayBetween(Date low, Date high) {
-        ObservableList<TamTruModel> listTamTru = FXCollections.observableArrayList();
+
+
+    /**
+     *
+     * @param low
+     * @param high
+     * @return thông tin tạm trú với ngày kết thúc tạm trú trong khoảng từ low dến high
+     */
+    public ObservableList<TamTruDisplayModel> getDisplayTamTruWhereDenNgayBetween(java.util.Date low, Date high){
+        ObservableList<TamTruDisplayModel> list= FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
-        String query = "select * from tam_tru where denNgay between ? and ?";
+        String query= "select CCCD, tam_tru.hoTen, ho_khau.diaChi as noiTamTru, tuNgay, denNgay, lydo, tam_tru.tinhTrang, tam_tru.idNguoiThucHien" +
+                "from tam_tru, ho_khau where tam_tru.maHoKhau= ho_khau.maHoKhau and denNgay between ? and ? ";
         try {
             PreparedStatement statement = connection.prepareStatement(query);
             statement.setDate(1, new java.sql.Date(low.getTime()));
@@ -177,114 +257,106 @@ public class TamTruService {
             e.printStackTrace();
         }
         try {
-            Statement statement = connection.createStatement();
+            Statement statement =connection.createStatement();
             ResultSet rs = statement.executeQuery(query);
             while (rs.next()) {
-                TamTruModel temp = new TamTruModel(
-                        rs.getInt("maTamTru"),
+                TamTruDisplayModel temp = new TamTruDisplayModel(
                         rs.getString("CCCD"),
                         rs.getNString("hoTen"),
                         rs.getNString("noiTamTru"),
                         rs.getDate("tuNgay"),
                         rs.getDate("denNgay"),
-                        rs.getString("lydo"),
+                        rs.getNString("lydo"),
                         rs.getNString("tinhTrang"),
                         rs.getInt("idNguoiThucHien"));
-                listTamTru.add(temp);
+
+                list.add(temp);
             }
             statement.close();
             connection.close();
-        } catch (Exception e) {
+
+        }catch (Exception e){
             e.printStackTrace();
         }
-
-        return listTamTru;
+        return list;
     }
 
-    public ObservableList<TamTruModel> getTamTruByTinhTrang(String tinhTrang) {
-        ObservableList<TamTruModel> listTamTru = FXCollections.observableArrayList();
+    /**
+     *
+     * @param tinhTrang
+     * @return thông tin tạm trú theo tình trạng
+     */
+    public ObservableList<TamTruDisplayModel> getDisplayTamTruByTinhTrang(String tinhTrang){
+        ObservableList<TamTruDisplayModel> list= FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
-        String query = "select * from tam_tru where tinhTrang = N'" + tinhTrang + "'";
+        String query= "select CCCD, tam_tru.hoTen, ho_khau.diaChi as noiTamTru, tuNgay, denNgay, lydo, tam_tru.tinhTrang, tam_tru.idNguoiThucHien" +
+                "from tam_tru, ho_khau where tam_tru.maHoKhau= ho_khau.maHoKhau and tinhTrang = N'"+ tinhTrang + "'";
         try {
-            Statement statement = connection.createStatement();
+            Statement statement =connection.createStatement();
             ResultSet rs = statement.executeQuery(query);
             while (rs.next()) {
-                TamTruModel temp = new TamTruModel(
-                        rs.getInt("maTamTru"),
+                TamTruDisplayModel temp = new TamTruDisplayModel(
                         rs.getString("CCCD"),
                         rs.getNString("hoTen"),
                         rs.getNString("noiTamTru"),
                         rs.getDate("tuNgay"),
                         rs.getDate("denNgay"),
-                        rs.getString("lydo"),
+                        rs.getNString("lydo"),
                         rs.getNString("tinhTrang"),
                         rs.getInt("idNguoiThucHien"));
-                listTamTru.add(temp);
+
+                list.add(temp);
             }
             statement.close();
             connection.close();
-        } catch (Exception e) {
+
+        }catch (Exception e){
             e.printStackTrace();
         }
-        return listTamTru;
+        return list;
     }
 
-    public ObservableList<TamTruModel> getTamTruByidNguoiThucHien(int idNguoiThucHien) {
-        ObservableList<TamTruModel> listTamTru = FXCollections.observableArrayList();
+    /**
+     *
+     * @param id
+     * @return thông tin tạm trú theo id người thực hiện
+     */
+    public ObservableList<TamTruDisplayModel> getDisplayTamTruByIDNguoiThucHien(int id){
+        ObservableList<TamTruDisplayModel> list= FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
-        String query = "select * from tam_tru where idNguoiThucHien = '" + idNguoiThucHien + "'";
+        String query= "select CCCD, tam_tru.hoTen, ho_khau.diaChi as noiTamTru, tuNgay, denNgay, lydo, tam_tru.tinhTrang, tam_tru.idNguoiThucHien" +
+                "from tam_tru, ho_khau where tam_tru.maHoKhau= ho_khau.maHoKhau and tam_tru.idNguoiThucHien = '"+ id + "'";
         try {
-            Statement statement = connection.createStatement();
+            Statement statement =connection.createStatement();
             ResultSet rs = statement.executeQuery(query);
             while (rs.next()) {
-                TamTruModel temp = new TamTruModel(
-                        rs.getInt("maTamTru"),
+                TamTruDisplayModel temp = new TamTruDisplayModel(
                         rs.getString("CCCD"),
                         rs.getNString("hoTen"),
                         rs.getNString("noiTamTru"),
                         rs.getDate("tuNgay"),
                         rs.getDate("denNgay"),
-                        rs.getString("lydo"),
+                        rs.getNString("lydo"),
                         rs.getNString("tinhTrang"),
                         rs.getInt("idNguoiThucHien"));
-                listTamTru.add(temp);
+
+                list.add(temp);
             }
             statement.close();
             connection.close();
-        } catch (Exception e) {
+
+        }catch (Exception e){
             e.printStackTrace();
         }
-        return listTamTru;
+        return list;
     }
 
-    public Optional<TamTruModel> getTamTruByMaTamTru(String maTamTru) {
-        Optional<TamTruModel> tamTruModel = Optional.empty();
-        Connection connection = ConnectionDatabase.getConnection();
-        String query = "select * from tam_tru where maTamTru = '" + maTamTru + "'";
-        try {
-            Statement statement = connection.createStatement();
-            ResultSet rs = statement.executeQuery(query);
-            if (rs.next()) {
-                TamTruModel temp = new TamTruModel(
-                        rs.getInt("maTamTru"),
-                        rs.getString("CCCD"),
-                        rs.getNString("hoTen"),
-                        rs.getNString("noiTamTru"),
-                        rs.getDate("tuNgay"),
-                        rs.getDate("denNgay"),
-                        rs.getString("lydo"),
-                        rs.getNString("tinhTrang"),
-                        rs.getInt("idNguoiThucHien"));
-                tamTruModel = Optional.of(temp);
-            }
-            statement.close();
-            connection.close();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return tamTruModel;
-    }
 
+    /**
+     *
+     * @param tamTruModel
+     * @return thêm một bản ghi tạm trú
+     */
     public boolean addTamTru(TamTruModel tamTruModel) {
         Connection connection = ConnectionDatabase.getConnection();
         String query = "insert into tam_tru(maHoKhau, CCCD, hoTen, tuNgay, denNgay, lydo, tinhTrang, idNguoiThucHien)" +
@@ -309,30 +381,11 @@ public class TamTruService {
         }
     }
 
-    public boolean addTamtru(TamTruDisplayModel tamTruDisplayModel) {
-        Connection connection = ConnectionDatabase.getConnection();
-        String query = "insert into tam_tru(maHoKhau, CCCD, hoTen, tuNgay, denNgay, lydo, tinhTrang, idNguoiThucHien)" +
-                "values (?, ?, ?, ?, ?, ?, ?, ?)";
-        try {
-            PreparedStatement statement = connection.prepareStatement(query);
-            statement.setString(1, new HoKhauService().getMaHoKhauByDiaChi(tamTruDisplayModel.getNoiTamTru()));
-            statement.setString(2, tamTruDisplayModel.getCCCD());
-            statement.setNString(3, tamTruDisplayModel.getHoTen());
-            statement.setDate(4, new java.sql.Date(tamTruDisplayModel.getTuNgay().getTime()));
-            statement.setDate(5, new java.sql.Date(tamTruDisplayModel.getDenNgay().getTime()));
-            statement.setNString(6, tamTruDisplayModel.getLyDo());
-            statement.setNString(7, tamTruDisplayModel.getTinhTrang());
-            statement.setInt(8, tamTruDisplayModel.getIdNguoiThucHien());
-            statement.executeUpdate();
-            statement.close();
-            connection.close();
-            return true;
-        } catch (Exception e) {
-            e.printStackTrace();
-            return false;
-        }
-    }
-
+    /**
+     *
+     * @param tamTruModel
+     * @return cập nhật một bản ghi tạm trú
+     */
 
     public boolean updateTamTru(TamTruModel tamTruModel) {
         Connection connection = ConnectionDatabase.getConnection();
@@ -366,6 +419,11 @@ public class TamTruService {
         }
     }
 
+    /**
+     *
+     * @param tamTruModel
+     * @return xóa một bản ghi tạm trú
+     */
     public boolean deleteTamTru(TamTruModel tamTruModel) {
         Connection connection = ConnectionDatabase.getConnection();
         String query = "delete from tam_tru where maTamTru = '" + tamTruModel.getMaTamTru() + "'";
@@ -380,6 +438,11 @@ public class TamTruService {
             return false;
         }
     }
+
+    /**
+     *
+     * @return số lượng bản ghi tạm trú
+     */
 
     public int getTamTruCount() {
         Connection connection = ConnectionDatabase.getConnection();
