+++ src/main/java/com/se07/controller/services/TamVangService.java
@@ -10,7 +10,15 @@ import java.sql.*;
 import java.util.Date;
 import java.util.Optional;
 
+/**
+ * class thực hiện các phương thức CRUD với tam_vang
+ */
 public class TamVangService {
+
+    /**
+     *
+     * @return tất cả bản ghi tạm vắng dưới dạng TamVangModel
+     */
     public ObservableList<TamVangModel> getAllTamVang() {
         ObservableList<TamVangModel> listTamVang = FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
@@ -38,91 +46,120 @@ public class TamVangService {
         return listTamVang;
     }
 
-    public ObservableList<TamVangModel> getTamVangByMaNhanKhau(String maNhanKhau) {
-        ObservableList<TamVangModel> listTamVang = FXCollections.observableArrayList();
+    /**
+     *
+     * @return tất cả bản ghi tạm vắng dưới dạng TamVangDisplayModel
+     */
+    public ObservableList<TamVangDisplayModel> getDisplayTamVang() {
+        ObservableList<TamVangDisplayModel> list = FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
-        String query = "select * from tam_vang where maNhanKhau = '" + maNhanKhau + "'";
+        String query = "select nhan_khau.hoTen, tam_vang.noiTamVang, , tuNgay, denNgay, lydo, tam_vang.tinhTrang, tam_vang.idNguoiThucHien" +
+                "from tam_vang, nhan_khau where tam_vang.maNhanKhau= nhan_khau.maNhanKhau";
         try {
             Statement statement = connection.createStatement();
             ResultSet rs = statement.executeQuery(query);
             while (rs.next()) {
-                TamVangModel temp = new TamVangModel(
-                        rs.getInt("maTamVang"),
-                        rs.getString("maNhanKhau"),
+                TamVangDisplayModel temp = new TamVangDisplayModel(
+                        rs.getNString("hoTen"),
                         rs.getNString("noiTamVang"),
                         rs.getDate("tuNgay"),
                         rs.getDate("denNgay"),
-                        rs.getString("lydo"),
+                        rs.getNString("lydo"),
                         rs.getNString("tinhTrang"),
                         rs.getInt("idNguoiThucHien"));
-                listTamVang.add(temp);
+
+                list.add(temp);
             }
             statement.close();
             connection.close();
+
         } catch (Exception e) {
             e.printStackTrace();
         }
-        return listTamVang;
+        return list;
     }
 
-    public ObservableList<TamVangModel> getTamVangByNoiTamVang(String noiTamVang) {
-        ObservableList<TamVangModel> listTamVang = FXCollections.observableArrayList();
+
+    /**
+     *
+     * @param hoTen
+     * @return thông tin tạm vắng theo họ tên
+     */
+    public ObservableList<TamVangDisplayModel> getDisplayTamVangByHoTen(String hoTen) {
+        ObservableList<TamVangDisplayModel> list = FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
-        String query = "select * from tam_vang where maNhanKhau like N'%" + noiTamVang + "%'";
+        String query = "select nhan_khau.hoTen, tam_vang.noiTamVang, , tuNgay, denNgay, lydo, tam_vang.tinhTrang, tam_vang.idNguoiThucHien" +
+                "from tam_vang, nhan_khau where tam_vang.maNhanKhau= nhan_khau.maNhanKhau and hoTen LIKE N'%"+ hoTen + "%'";
         try {
             Statement statement = connection.createStatement();
             ResultSet rs = statement.executeQuery(query);
             while (rs.next()) {
-                TamVangModel temp = new TamVangModel(
-                        rs.getInt("maTamVang"),
-                        rs.getString("maNhanKhau"),
+                TamVangDisplayModel temp = new TamVangDisplayModel(
+                        rs.getNString("hoTen"),
                         rs.getNString("noiTamVang"),
                         rs.getDate("tuNgay"),
                         rs.getDate("denNgay"),
-                        rs.getString("lydo"),
+                        rs.getNString("lydo"),
                         rs.getNString("tinhTrang"),
                         rs.getInt("idNguoiThucHien"));
-                listTamVang.add(temp);
+
+                list.add(temp);
             }
             statement.close();
             connection.close();
+
         } catch (Exception e) {
             e.printStackTrace();
         }
-        return listTamVang;
+        return list;
     }
 
-    public ObservableList<TamVangModel> getTamVangBytinhTrang(String tinhTrang) {
-        ObservableList<TamVangModel> listTamVang = FXCollections.observableArrayList();
+    /**
+     *
+     * @param noiTamVang
+     * @return thông tin tạm vắng theo nơi tạm vắng
+     */
+    public ObservableList<TamVangDisplayModel> getDisplayTamVangByNoiTamVang(String noiTamVang) {
+        ObservableList<TamVangDisplayModel> list = FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
-        String query = "select * from tam_vang where tinhTrang = '" + tinhTrang + "'";
+        String query = "select nhan_khau.hoTen, tam_vang.noiTamVang, , tuNgay, denNgay, lydo, tam_vang.tinhTrang, tam_vang.idNguoiThucHien" +
+                "from tam_vang, nhan_khau where tam_vang.maNhanKhau= nhan_khau.maNhanKhau and" +
+                " noiTamVang LIKE N'%"+ noiTamVang + "%'";
         try {
             Statement statement = connection.createStatement();
             ResultSet rs = statement.executeQuery(query);
             while (rs.next()) {
-                TamVangModel temp = new TamVangModel(
-                        rs.getInt("maTamVang"),
-                        rs.getString("maNhanKhau"),
+                TamVangDisplayModel temp = new TamVangDisplayModel(
+                        rs.getNString("hoTen"),
                         rs.getNString("noiTamVang"),
                         rs.getDate("tuNgay"),
                         rs.getDate("denNgay"),
-                        rs.getString("lydo"),
+                        rs.getNString("lydo"),
                         rs.getNString("tinhTrang"),
                         rs.getInt("idNguoiThucHien"));
-                listTamVang.add(temp);
+
+                list.add(temp);
             }
             statement.close();
             connection.close();
+
         } catch (Exception e) {
             e.printStackTrace();
         }
-        return listTamVang;
+        return list;
     }
 
-    public ObservableList<TamVangModel> getTamVangWhereTuNgayBetween(Date low, Date high) {
-        ObservableList<TamVangModel> listTamVang = FXCollections.observableArrayList();
+    /**
+     *
+     * @param low
+     * @param high
+     * @return thông tin tạm vắng với ngày bắt đầu tạm vắng nằm giữa low và high
+     */
+    public ObservableList<TamVangDisplayModel> getDisplayTamVangWhereTuNgayBetween(Date low, Date high) {
+        ObservableList<TamVangDisplayModel> list = FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
-        String query = "select * from tam_vang where tuNgay between ? and ?";
+        String query = "select nhan_khau.hoTen, tam_vang.noiTamVang, , tuNgay, denNgay, lydo, tam_vang.tinhTrang, tam_vang.idNguoiThucHien" +
+                "from tam_vang, nhan_khau where tam_vang.maNhanKhau= nhan_khau.maNhanKhau and tuNgay between ? and ?";
         try {
             PreparedStatement statement = connection.prepareStatement(query);
             statement.setDate(1, new java.sql.Date(low.getTime()));
@@ -133,34 +170,44 @@ public class TamVangService {
         } catch (Exception e) {
             e.printStackTrace();
         }
+
         try {
             Statement statement = connection.createStatement();
             ResultSet rs = statement.executeQuery(query);
             while (rs.next()) {
-                TamVangModel temp = new TamVangModel(
-                        rs.getInt("maTamVang"),
-                        rs.getString("maNhanKhau"),
+                TamVangDisplayModel temp = new TamVangDisplayModel(
+                        rs.getNString("hoTen"),
                         rs.getNString("noiTamVang"),
                         rs.getDate("tuNgay"),
                         rs.getDate("denNgay"),
-                        rs.getString("lydo"),
+                        rs.getNString("lydo"),
                         rs.getNString("tinhTrang"),
                         rs.getInt("idNguoiThucHien"));
-                listTamVang.add(temp);
+
+                list.add(temp);
             }
             statement.close();
             connection.close();
+
         } catch (Exception e) {
             e.printStackTrace();
         }
+        return list;
 
-        return listTamVang;
+
     }
 
-    public ObservableList<TamVangModel> getTamVangWhereDenNgayBetween(Date low, Date high) {
-        ObservableList<TamVangModel> listTamVang = FXCollections.observableArrayList();
+    /**
+     *
+     * @param low
+     * @param high
+     * @return thông tin tạm vắng với ngày kết thúc tạm vắng giữa ngày low và high
+     */
+    public ObservableList<TamVangDisplayModel> getDisplayTamVangWhereDenNgayBetween(Date low, Date high) {
+        ObservableList<TamVangDisplayModel> list = FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
-        String query = "select * from tam_vang where denNgay between ? and ?";
+        String query = "select nhan_khau.hoTen, tam_vang.noiTamVang, , tuNgay, denNgay, lydo, tam_vang.tinhTrang, tam_vang.idNguoiThucHien" +
+                "from tam_vang, nhan_khau where tam_vang.maNhanKhau= nhan_khau.maNhanKhau and denNgay between ? and ?";
         try {
             PreparedStatement statement = connection.prepareStatement(query);
             statement.setDate(1, new java.sql.Date(low.getTime()));
@@ -171,57 +218,110 @@ public class TamVangService {
         } catch (Exception e) {
             e.printStackTrace();
         }
+
         try {
             Statement statement = connection.createStatement();
             ResultSet rs = statement.executeQuery(query);
             while (rs.next()) {
-                TamVangModel temp = new TamVangModel(
-                        rs.getInt("maTamVang"),
-                        rs.getString("maNhanKhau"),
+                TamVangDisplayModel temp = new TamVangDisplayModel(
+                        rs.getNString("hoTen"),
                         rs.getNString("noiTamVang"),
                         rs.getDate("tuNgay"),
                         rs.getDate("denNgay"),
-                        rs.getString("lydo"),
+                        rs.getNString("lydo"),
                         rs.getNString("tinhTrang"),
                         rs.getInt("idNguoiThucHien"));
-                listTamVang.add(temp);
+
+                list.add(temp);
             }
             statement.close();
             connection.close();
+
         } catch (Exception e) {
             e.printStackTrace();
         }
+        return list;
 
-        return listTamVang;
+
     }
 
-    public ObservableList<TamVangModel> getTamVangByidNguoiThucHien(int id) {
-        ObservableList<TamVangModel> listTamVang = FXCollections.observableArrayList();
+    /**
+     *
+     * @param tinhTrang
+     * @return thông tin tạm vắng theo tình trạng
+     */
+    public ObservableList<TamVangDisplayModel> getDisplayTamVangByTinhTrang(String tinhTrang) {
+        ObservableList<TamVangDisplayModel> list = FXCollections.observableArrayList();
         Connection connection = ConnectionDatabase.getConnection();
-        String query = "select * from tam_vang where idNguoiThucHien = '" + id + "'";
+        String query = "select nhan_khau.hoTen, tam_vang.noiTamVang, , tuNgay, denNgay, lydo, tam_vang.tinhTrang, tam_vang.idNguoiThucHien" +
+                "from tam_vang, nhan_khau where tam_vang.maNhanKhau= nhan_khau.maNhanKhau and" +
+                " tinhTrang LIKE N'"+ tinhTrang + "'";
         try {
             Statement statement = connection.createStatement();
             ResultSet rs = statement.executeQuery(query);
             while (rs.next()) {
-                TamVangModel temp = new TamVangModel(
-                        rs.getInt("maTamVang"),
-                        rs.getString("maNhanKhau"),
+                TamVangDisplayModel temp = new TamVangDisplayModel(
+                        rs.getNString("hoTen"),
                         rs.getNString("noiTamVang"),
                         rs.getDate("tuNgay"),
                         rs.getDate("denNgay"),
-                        rs.getString("lydo"),
+                        rs.getNString("lydo"),
                         rs.getNString("tinhTrang"),
                         rs.getInt("idNguoiThucHien"));
-                listTamVang.add(temp);
+
+                list.add(temp);
             }
             statement.close();
             connection.close();
+
         } catch (Exception e) {
             e.printStackTrace();
         }
-        return listTamVang;
+        return list;
     }
 
+
+    /**
+     *
+     * @param id
+     * @return thông tin tạm vắng theo id người thực hiện
+     */
+    public ObservableList<TamVangDisplayModel> getDisplayTamVangByIDNguoiThucHien(int id) {
+        ObservableList<TamVangDisplayModel> list = FXCollections.observableArrayList();
+        Connection connection = ConnectionDatabase.getConnection();
+        String query = "select nhan_khau.hoTen, tam_vang.noiTamVang, , tuNgay, denNgay, lydo, tam_vang.tinhTrang, tam_vang.idNguoiThucHien" +
+                "from tam_vang, nhan_khau where tam_vang.maNhanKhau= nhan_khau.maNhanKhau and" +
+                " idNguoiThucHien ='"+ id + "'";
+        try {
+            Statement statement = connection.createStatement();
+            ResultSet rs = statement.executeQuery(query);
+            while (rs.next()) {
+                TamVangDisplayModel temp = new TamVangDisplayModel(
+                        rs.getNString("hoTen"),
+                        rs.getNString("noiTamVang"),
+                        rs.getDate("tuNgay"),
+                        rs.getDate("denNgay"),
+                        rs.getNString("lydo"),
+                        rs.getNString("tinhTrang"),
+                        rs.getInt("idNguoiThucHien"));
+
+                list.add(temp);
+            }
+            statement.close();
+            connection.close();
+
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return list;
+    }
+
+
+    /**
+     *
+     * @param maTamVang
+     * @return thông tin tạm vắng dưới dạng TamVangModel theo mã tạm vắng
+     */
     public Optional<TamVangModel> getTamVangByMaTamVang(String maTamVang) {
         Optional<TamVangModel> tamVangModel = Optional.empty();
         Connection connection = ConnectionDatabase.getConnection();
@@ -249,6 +349,12 @@ public class TamVangService {
         return tamVangModel;
     }
 
+    /**
+     *
+     * @param tamVangModel
+     * @return thêm vào cơ sở dữ liệu một bản ghi tạm vắng
+     */
+
     public boolean addTamVang(TamVangModel tamVangModel) {
         Connection connection = ConnectionDatabase.getConnection();
         String query = "insert into tam_vang(maNhanKhau, noiTamVang, tuNgay, denNgay, lydo, " +
@@ -271,6 +377,12 @@ public class TamVangService {
             return false;
         }
     }
+
+    /**
+     *
+     * @param tamVangModel
+     * @return cập nhật thông tin bản ghi tạm vắng
+     */
 
     public boolean updateTamVang(TamVangModel tamVangModel) {
         Connection connection = ConnectionDatabase.getConnection();
@@ -302,6 +414,12 @@ public class TamVangService {
         }
     }
 
+    /**
+     *
+     * @param tamVangModel
+     * @return xóa một bản ghi tạm vắng
+     */
+
     public boolean deleteTamVang(TamVangModel tamVangModel) {
         Connection connection = ConnectionDatabase.getConnection();
         String query = "delete from tam_vang where maTamVang = '" + tamVangModel.getMaTamVang() + "'";
@@ -317,6 +435,10 @@ public class TamVangService {
         }
     }
 
+    /**
+     *
+     * @return số lượng bản ghi tạm vắng
+     */
     public int getTamVangCount() {
         Connection connection = ConnectionDatabase.getConnection();
         String query = "select count(*) from tam_vang where tinhTrang = N'Đã xác nhận'";
